clang fix 1

+++ tasks/yusupkina_m_elem_vec_sum/common/include/common.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <tuple>
 #include <vector>
@@ -9,7 +10,7 @@
 namespace yusupkina_m_elem_vec_sum {
 
 using InType = std::vector<int>;
-using OutType = long long;
+using OutType = int64_t;
 using TestType = std::tuple<int, std::string>;
 using BaseTask = ppc::task::Task<InType, OutType>;
